+++ include/nfx/Gl.h
@@ -9,6 +9,8 @@
 
 #include "nfx/graphics/gl/core/buffers/Buffer.h"
 #include "nfx/graphics/gl/core/buffers/VertexArray.h"
+#include "nfx/graphics/gl/core/framebuffers/Framebuffer.h"
+#include "nfx/graphics/gl/core/framebuffers/Renderbuffer.h"
 #include "nfx/graphics/gl/core/shaders/ShaderProgram.h"
 #include "nfx/graphics/gl/core/textures/Texture2D.h"
 #include "nfx/graphics/gl/core/Context.h"